updated get_program, made many new commands

+++ main.cpp
@@ -5,7 +5,7 @@ using namespace CPULib;
 
 int main()
 {
-    cpu<int> cpu_1;
+    cpu<long long> cpu_1;
     cpu_1.get_program();
     cpu_1.start();
